تلاع

+++ resources/views/admin/categories/index.blade.php
@@ -87,8 +87,10 @@
                                 @if(Auth::user()->update_data)
                                     <td class="text-center blue-color" ><a href="/admin-panel/categories/edit/{{ $category->id }}" ><i class="far fa-edit"></i></a></td>
                                 @endif
-                                @if(Auth::user()->delete_data)
-                                    <td class="text-center blue-color" ><a onclick="return confirm('Are you sure you want to delete this item?');" href="/admin-panel/categories/delete/{{ $category->id }}" ><i class="far fa-trash-alt"></i></a></td>
+                                @if($category->id != 7)
+                                    @if(Auth::user()->delete_data)
+                                        <td class="text-center blue-color" ><a onclick="return confirm('Are you sure you want to delete this item?');" href="/admin-panel/categories/delete/{{ $category->id }}" ><i class="far fa-trash-alt"></i></a></td>
+                                    @endif
                                 @endif
                                 <?php $i++; ?>
                             </tr>
